Tweak message list to have 20px tall messages
This makes the UI somewhat consistent, but it could be reviewed at some
point, how it all should fit together.

Perhaps, messages should be 50% height of channel list entries, which
might not be practical. It seems at least, that the UI looks more
unified with similarly sized list items, for the moment.